Updated and improved the script. See description:
Now the code works again.
All the songs are put in a subfolder whose name is the playlist name.
Video files used to make the conversion are delete upon use.
Generated files now have a better quality.
The script doesn't ask you anymore for the folder, it simply works from the folder where the script is executed.